Decrease doc id max to 32bit for peformance reasons

+++ index.cpp
@@ -14,7 +14,7 @@ const char *usage = "\
 Usage: index [file]\
 ";
 
-void index_write(char const *filename, char *buffer, dynamic_array<std::pair<char *, size_t>> *docNos, hash_table<posting, size_t> *dictionary) {
+void index_write(char const *filename, char *buffer, dynamic_array<std::pair<char *, size_t>> *docNos, hash_table<posting, uint32_t> *dictionary) {
 	FILE *fh = fopen(filename, "w");
 	if (fh == NULL) {
 		fprintf(stderr, "ERROR: Failed to open index.dat for writing\n");
@@ -54,8 +54,8 @@ int main(int argc, char **argv) {
 	enum token_type token;
 
 	dynamic_array<std::pair<char *, size_t>> *docNos = new dynamic_array<std::pair<char *, size_t>>();
-	hash_table<posting, size_t> *dictionary = new hash_table<posting, size_t>();
-	size_t docI = 0;
+	hash_table<posting, uint32_t> *dictionary = new hash_table<posting, uint32_t>();
+	uint32_t docI = 0;
 	do {
 		token = tok->next(tok_buffer);
 		if (token == DOCNO) {
